public visibility for getBy
I was looking for a lib that get a country code by country name and found this one. Unfortunately, it only works from a country code to a country name. I think it could be great to search by anything. That's why I changed the visibility. One other way may be to add a `getByName` method but I'm not sure it's better.